Version 1
This version takes all the new stops for a route and finds which current stops they are the closest to and then automatically inserts the new stop right after that current location in the dataframe. The final result is a new route with the new stops added (in a way that optimizes the route duration).

Problem: There are cases where the new stop is closest to a current location and it is therefore put right after it in the new route but sometimes, the new stop is actually on the way to the current location not right after. I have to answer the question: How can I find if the new stop is before or after the current location? (How can I find if the new stop is between the previous location and the current closest location?)