Add Data
This is the demo data used to test the code for version 1. It brings up a problem: what happens if the newStop is not closest to any current stops but the last one? np.insert replaces stops not adds stops so I added an if statement to account for the addition of the new stop to the end of the route if that's where it belongs.